refactored project to best practices (#32)
