+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	kvs "github.com/defoeam/kvs/kv"
+	kvs "github.com/defoeam/kvs/internal"
 )
 
 func main() {
